Remove ThreatAction, ThreatStore, ThreatRestClient
Scan API responds with a site's threats within the scan state response.
With the current API structure, "Fetch Threat" and "Fetch Threats" APIs are not required by the app, which might have wanted a dedicated ThreatStore.

To keep things simple, we're deleting the store and rest client for threats and retaining just the ThreatModel, ThreatMapper and corresponding custom serializers.

Scan module can now have separate tables for storing scan state and threats and need not interact with a separate ThreatStore.