Analysis: Suppress member name equals class name warning for wc model
Description: "This rule reports a member that has the same as the
containing class or object. This might result in confusion. The member
should either be renamed or changed to a constructor. Factory functions
that create an instance of the class are exempt from this rule."

For more info see: https://detekt.dev/docs/rules/naming
#membernameequalsclassname