Analysis: Resolve new api lint warnings for html from html static call
Replacing 'Html.fromHtml(html)' with the
'HtmlCompat.fromHtml(html, HtmlCompat.FROM_HTML_MODE_LEGACY)' compat
related version is the recommended action to resolve this kind of Lint
warnings.

However, note that this was only possible for 'priceAmountHtmlTag'.
Doing the same for the other two, 'priceAmountHtmlTag' and
'itemHtmlTag', isn't possible as those fields can end up being nullable,
that is, due to the fact that the 'LeaderboardItemRow.display'
constructor argument is nullable as well. As such, and because
'HtmlCompat.fromHtml(html, HtmlCompat.FROM_HTML_MODE_LEGACY)' accepts a
non-null String 'source', it is not possible to resolve these two Lint
warnings.