Deps: Remove unnecessary org apache commons lang dependency for woo
This dependency was added as part of the
000e42fc535a7591ffed5657aaf28d9ecd7c8fa2 commit on the
'plugins:woocommerce' module, most probably in order to be later used as
part of the 148f7fd9a177da3f84d3bd60a09d74cc99049c13 commit, but this
time, on the 'example' module on 'WCOrderListActivity' class in order to
import 'org.apache.commons.lang3.time.DateUtils'.

However, the 'example' module already has such a
'org.apache.commons:commons-lang3' dependency declaration. As such, this
similar dependency declaration on the 'plugins:woocommerce' module is
totally unused and can be safely removed.