Analysis: Remove unnecessary lint configuration for example module
It doesn't seem that this Lint related warning configuration for the
'InvalidPackage' check is necessary for the 'example' module. It is
better to remove that and have Lint being configured as straightforward
as possible.

FYI: This unnecessary Lint configuration was added as part of this very
first commit on this repo back on October 27th, 2015:
877745651bfeddeab7dea4d435d5a5230d6e99f6