Analysis: Add missing n-a to on taxonomy changed
FYI: 'n-a' stands for 'nullability annotations'.

This change makes the 'causeOfChange' field on 'OnTaxonomyChanged'
non-null, which in its turn caused few updates in terms of how
'OnTaxonomyChanged' is being constructed. Otherwise, this
'causeOfChange' field couldn't have been non-null, but having that as
nullable is not justified as it can be easily become and used as
non-null.