Refactor: Remove unnecessary taxonomy name field from taxonomy changed
This 'taxonomyName' was only used within the 'example' app, just for
logging purposes. As such, production wise it seems to not be holding
any real value. This is especially true since JP/WPAndroid is not using
this field anywhere. Thus, it is safe to remove this field and not
having to deal with a nullable 'taxonomyName' field for no particular
good reason, and thus making 'OnTaxonomyChanged' much more
straightforward as a class.