Analysis: Add missing n-a to term model
FYI: 'n-a' stands for 'nullability annotations'.

1) A 'TermModel' is currently used either for a 'category' or a 'tag'.
2) When a new 'category' is initialized, it always has a 'mTaxonomy',
which is actually a 'category', and also a 'mName', the category name
which is given by the user. However, it doesn't have a 'mSlug' or a
'mDescription'. The 'mSlug' gets added to 'TermModel' with the API
response. A 'category' has no description, thus 'mDescription' is either
'null' or being transformed to an empty string ("") with the API
response.
3) When a new 'tag' is initialized, it always has a 'mTaxonomy', which
is actually a 'tag', and also a 'nName', the tag name, which is given by
the user. In contrast to a 'category' type 'TermModel', a 'tag' does
have a 'mSlug' and can have a 'mDescription' too. The 'mSlug' gets added
during creation via 'mTerm.setSlug(sanitizeWithDashes(thisName))' and as
such getting its 'mSlug' from its 'mName'. The 'mDescription' is either
and empty string (""), if the user chooses to omit this step, or any
other string description, which is given by the user.

PS: This 'NotNullFieldNotInitialized' warning got suppressed because a
table related 'model' class can never have its fields initialized via a
constructor initialization, or otherwise for that matter.