Cleanup: Remove unused transient level field from comment model
This transient 'level' field doesn't seem to be used within FluxC or
any client app that might be using the this 'CommentModel' model.

FYI: There is another such definition of a transient 'level' field,
which is actually still in use, but that is on the JP/WPAndroid level
and via the 'ReaderComment' model.

PS: Actually, JP/WPAndroid is the only client that is using this
'CommentModel' model, or 'CommentStore' for that matter.