Test: Change allow nulls parameter with with empty on cmts mapper test
Also, this 'SameParameterValue' warning got suppressed for the
'getDefaultCommentList(...)' function as it is currently being used once
and as 'getDefaultCommentList(false)'. How this doesn't mean it will not
be used in the future, or wasn't used in the past. As such, keeping this
method signature and suppressing this warning for the time being seems
to be the better approach here.

Related PR Comment: https://github.com/wordpress-mobile/
WordPress-FluxC-Android/pull/2868#discussion_r1362018044