Analysis: Add missing n-a to get images/videos/audio/docs on media store
FYI: 'n-a' stands for 'nullability annotations'

PS: Also, as part of this change, the 'getSiteImageCount(...)' method
got removed as it is only being used within 'MediaStoreTest'. The is
no reason to create a new method for 'getSiteImages(siteModel).size()'
alone, and more so, when no client apps like JP/WP/WCAndroid are using
it.