Analysis: Add missing n-a to match site media related on media store
FYI: 'n-a' stands for 'nullability annotations'

PS: Also, as part of this change, the 'getUnattachedSiteMediaCount(...)'
method got removed as it is only being used within 'MediaStoreTest'. The
is no reason to have a new method for 'getSiteImages(siteModel).size()'
alone, and more so, when no client apps like JP/WP/WCAndroid are using
it.