Build: Update dependency analysis configuration on fluxc processor
All 'unused' related advises for the 'fluxc-processor' are incorrect
(for an unknown to me reason) and as such those were instead suppressed.
The 'onUnusedDependencies', 'onUnusedAnnotationProcessors' and
'onRedundantPlugins' exclusion configuration were used to achieve that.

------------------------------------------------------------------------

Advice for :fluxc-processor
Unused dependencies which should be removed:
  api 'org.jetbrains.kotlin:kotlin-stdlib:1.9.25'
  implementation sharedLibs.google.autoService

Unused annotation processors that should be removed:
  kapt sharedLibs.google.autoService

Unused plugins that can be removed:
  kotlin-kapt: this project has the kotlin-kapt
   (org.jetbrains.kotlin.kapt) plugin applied, but there are no used
   annotation processors.