Build: Update jvm target & source/target compatibility to java 11
Although this change is not required, it is high time it is done in
order to update the Java target and compatibility across the project and
thus be prepared for any future updates.

------------------------------------------------------------------------

FYI: For example, on JP/WPAndroid this change was actually necessary
because otherwise updating to Kotlin 2.0 was failing the build with the
below:

FAILURE: Build failed with an exception.

* What went wrong:
Execution failed for task ':libs:processors:compileKotlin'.
> Could not resolve all files for configuration
   ':libs:processors:compileClasspath'.
   > Could not resolve com.google.devtools.ksp:symbol-processing-api:
      2.0.21-1.0.26.
     Required by:
         project :libs:processors
      > Dependency resolution is looking for a library compatible with
       JVM runtime version 8, but 'com.google.devtools.ksp:
       symbol-processing-api:2.0.21-1.0.26' is only compatible with JVM
       runtime version 11 or newer.

PR Commit: https://github.com/wordpress-mobile/WordPress-Android/pull/
21428/commits/6abe01ec983d0765d4bd1466030097109867ac56