Remove unused code
I've checked the history, and `CardReaderPaymentViewModel::reFetchOrder` has never been invoked outside of the class. When extracting payment flow to the controller, I missed the `@VisibleForTesting` annotation and treated it as a public function when it was supposed to be treated as `private`. I'm removing this method from VM – it's supposed to be invoked internally in `CardReaderPaymentController`.