Fix crash when camera not available
`ProcessCameraProvider::bindToLifecycle` is throwing IllegalStateException in case the requested camera is not available on the device. However, we were catching this exception outside the View.post {...} lambda in which the `ProcessCameraProvider::bindToLifecycle` was invoked asynchronously. This resulted in exception not being caught. The post method schedules the code to run on the main thread, but the try-catch block only surrounds the scheduling call, not the actual execution of the code inside the post block.