CI: Add restore cache and run on targeted pipeline related jobs only
PS: The targeted 'pipeline' related jobs are:
- Mobile App
- Wear App
- Lint
- Unit Tests
- Android tests

PS: The 'detekt' job isn't targeted because it only takes about 20+
seconds of network request to download all the dependencies needed for
this specific job, which is about the same time it take for the
dependency cache to actually be restored (10+ seconds). Thus, this
optimization is not really worth it for this specific job. For all those
other jobs targeted, it take more than a minute of network requests to
download all the dependencies needed, and as such worth the diff is
worth it.