Analysis: Resolve unsafe opt in usage lint error on google mlkit cs
Annotating the whole 'recogniseCode(...)' suspend function with
'@ExperimentalGetImage' instead of just doing that its inner
'MediaImageProvider.provideImage(...)' call, and then propagating this
annotation on that single usage of 'recogniseCode(...)', which is within
'BarcodeScanningViewModel.startCodesRecognition()', is all it takes to
properly resolve this issue.

Error Message: "This declaration is opt-in and its usage should be
marked with @androidx.camera.core.ExperimentalGetImage or
@OptIn(markerClass = androidx.camera.core.ExperimentalGetImage.class)"

Explanation: "This API has been flagged as opt-in with error-level severity.

Any declaration annotated with this marker is considered part of an
unstable or otherwise non-standard API surface and its call sites should
accept the opt-in aspect of it by using the @OptIn annotation, using the
marker annotation -- effectively causing further propagation of the
opt-in aspect -- or configuring the UnsafeOptInUsageError check's
options for project-wide opt-in.

To configure project-wide opt-in, specify the opt-in option value in
lint.xml as a comma-delimited list of opted-in annotations:
<lint>
    <issue id="UnsafeOptInUsageError">
        <option name="opt-in" value="com.foo.ExperimentalBarAnnotation"
         />
    </issue>
</lint>"