Deps: Stop ignoring the agp update
The AGP update no longer needs to be synced with other in-house
library and client projects. As such, there is no need for Dependabot to
ignore them anymore.

This limitation got lifted due to the fact that:
- 'FluxC' got vendored into this project, and thus the need to sync the
AGP version with JP/WPAndroid is no longer relevant (paaHJt-7GA-p2).
- It has been decided that any project that requires compatibility with
composite build on a library level, they will have to make the change
themselves, when and if needed (paaHJt-6Qv-p2).