Create `WooUpgradeRequiredDialogFragment`
This commit introduces the `WooUpgradeRequiredDialogFragment`, which is a custom DialogFragment used to display a dialog that informs the user about a required WooCommerce upgrade. This dialog is implemented using Jetpack Compose, and it allows the user to navigate to the WooCommerce upgrade instructions using a custom Chrome tab. It also includes an option to dismiss the dialog.